deploy/aws/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
pytest>=7.0.0
pytest-asyncio>=0.21.0
mypy>=0.950
black>=22.3.0
flake8>=4.0.1 